+++ kan/kan/__init__.py
@@ -1,0 +1,3 @@
+from .kan import KANLinear, KAN
+
+__all__ = ["KANLinear", "KAN"]
